+++ src/Jigoshop/Endpoint/DownloadFile.php
@@ -146,7 +146,7 @@ class DownloadFile implements Processable
                     default:
                         $type = 'application/force-download';
                 }
-                $this->wp->doAction('jigoshop\downloadable\before_download', $file, $order);
+                $this->wp->doAction('jigoshop\endpoint\downloadable\before_download', $file, $order);
                 @session_write_close();
                 @set_time_limit(0);
                 @ob_end_clean();