+++ src/Jigoshop/Frontend/Page/Cart.php
@@ -481,7 +481,7 @@ class Cart implements PageInterface
 							}
 						}
 
-						do_action('jigoshop\cart\before_checkout', $cart);
+						$this->wp->doAction('jigoshop\cart\before_checkout', $cart);
 
 						$this->cartService->save($cart);
 						$this->messages->preserveMessages();