+++ templates/user/account/orders/single.php
@@ -12,6 +12,7 @@ use Jigoshop\Helper\Render;
  * @var $myAccountUrl string URL to my account.
  * @var $listUrl string URL to orders list.
  * @var $showWithTax bool Whether to show product price with or without tax.
+ * @var $suffix string
  * @var $getTaxLabel \Closure Function to retrieve tax label.
  */
 ?>
@@ -59,7 +60,7 @@ use Jigoshop\Helper\Render;
 	</thead>
 	<tbody>
 		<?php foreach($order->getItems() as $key => $item): /** @var $item \Jigoshop\Entity\Order\Item */ ?>
-			<?php Render::output('shop/checkout/item/'.$item->getType(), array('cart' => $order, 'key' => $key, 'item' => $item, 'showWithTax' => $showWithTax)); ?>
+			<?php Render::output('shop/checkout/item/'.$item->getType(), array('cart' => $order, 'key' => $key, 'item' => $item, 'showWithTax' => $showWithTax, 'suffix' => $suffix)); ?>
 		<?php endforeach; ?>
 		<?php do_action('jigoshop\checkout\table_body', $order); ?>
 	</tbody>